+++ src/environment/mod.rs
@@ -39,12 +39,12 @@ pub fn get_range(coords: NodeRange) -> Range {
 
     Range {
         start: Position {
-            line: u32::from(start.0),
-            character: u32::from(start.1),
+            line: start.0,
+            character: start.1,
         },
         end: Position {
-            line: u32::from(end.0),
-            character: u32::from(end.1),
+            line: end.0,
+            character: end.1,
         },
     }
 }
